refactor: rename ema_data to ema_db
